pyproject.toml
[tool.poetry]
name = "core"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "^3.7"
aniso8601 = "8.0.0"
attrs = "19.3.0"
certifi = "2020.4.5.1"
cffi = "1.14.0"
chardet = "3.0.4"
Flask = "1.1.4"
Flask-Cors = "3.0.8"
Flask-Executor = "0.9.3"
Flask-JWT = "0.3.2"
flask-oidc = "1.4.0"
Flask-RESTful = "0.3.8"
flask-restx = "0.2.0"
gunicorn = "20.0.4"
itsdangerous = "1.1.0"
Jinja2 = "2.11.2"
jsonschema = "3.2.0"
MarkupSafe = "1.1.1"
psycopg2-binary = "2.8.6"
PyJWT = "1.4.2"
python-json-logger = "0.1.11"
pytz = "2020.1"
requests = "2.23.0"
SQLAlchemy = "1.3.17"
toml = "0.10.1"
urllib3 = "1.25.9"
Werkzeug = "1.0.1"
wrapt = "1.12.1"
flask-sqlalchemy = "2.4.4"
python-ldap = "3.3.1"
python-dotenv = "0.19.1"
pydantic = "1.8.2"
asyncpg = "0.25.0"
opentelemetry-exporter-jaeger = "1.6.2"
opentelemetry-instrumentation = "0.26b1"
opentelemetry-instrumentation-flask = "0.26b1"
opentelemetry-instrumentation-requests = "0.26b1"
minio = "7.1.2"
httpx = "0.22.0"
pilot-platform-common = "0.0.25"

[tool.poetry.dev-dependencies]
pytest = "6.0.2"
pytest-env = "0.6.2"
pytest-mock = "3.7.0"
coverage = "6.3.2"
requests-mock = "1.9.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
